+++ src/jdk.jfr/share/classes/jdk/jfr/internal/jfc/model/Utilities.java
@@ -22,7 +22,7 @@
  * or visit www.oracle.com if you need additional information or have any
  * questions.
  */
-package jdk.jfr.internal.parameters;
+package jdk.jfr.internal.jfc.model;
 
 import java.util.StringJoiner;
