+++ src/jdk.jfr/share/classes/jdk/jfr/internal/jfc/model/XmlAnd.java
@@ -22,34 +22,28 @@
  * or visit www.oracle.com if you need additional information or have any
  * questions.
  */
-package jdk.jfr.internal.parameters;
+package jdk.jfr.internal.jfc.model;
 
-import java.util.List;
-import java.util.Optional;
+// Corresponds to <and>
+final class XmlAnd extends XmlExpression {
 
-public abstract class XmlInput extends XmlElement implements XmlVariable {
-
-    public abstract String getOptionSyntax();
-
-    abstract void configure(UserInterface ui) throws AbortException;
-
-    abstract void configure(String value);
-
-    public final Optional<String> getContentType() {
-        return optional("contentType");
+    @Override
+    boolean isEntity() {
+        return false;
     }
 
     @Override
-    public final String getName() {
-        return attribute("name");
-    }
-
-    public final String getLabel() {
-        return attribute("label");
-    }
-
-    @Override
-    protected List<String> attributes() {
-        return List.of("name", "label");
+    protected Result evaluate() {
+        Result result = Result.NULL;
+        for (XmlElement e : getProducers()) {
+            Result r = e.evaluate();
+            if (r.isFalse()) {
+                return Result.FALSE;
+            }
+            if (r.isTrue()) {
+                result = Result.TRUE;
+            }
+        }
+        return result;
     }
 }
